feat: map props depending on platform
