+++ packages/image/props-to-state.js
@@ -1,5 +1,14 @@
-export default ({ source }) => ({
-  source: {
-    uri: source && source.uri
-  }
-});
+import { Dimensions } from "react-native";
+
+const window = Dimensions.get("window");
+export default ({ source }) => {
+  const uri = source.uri;
+
+  return {
+    height: 1,
+    source: {
+      uri: uri && uri.indexOf("//") === 0 ? `https:${uri}` : uri
+    },
+    width: window.width
+  };
+};
